swe/python/2023/async-k8s/requirements.txt
# pip install kubernetes_asyncio;
# pip freeze | tee requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.2.0
frozenlist==1.3.3
idna==3.4
kubernetes-asyncio==24.2.3
multidict==6.0.4
python-dateutil==2.8.2
PyYAML==6.0
six==1.16.0
urllib3==2.0.3
yarl==1.9.2
